refactor(#23): extract shared LogFileScreen and LogFileViewModel
